requirements.txt
tensorflow
tensorflow-datasets
imgaug
pillow
pydicom
pylibjpeg
pylibjpeg-libjpeg
numpy>1.19